requirements.txt
absl-py==1.1.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
cloudpickle==2.1.0
cycler==0.11.0
Cython==0.29.30
flatbuffers==1.12
fonttools==4.34.4
gast==0.4.0
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
gym==0.21.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.12.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
libclang==14.0.1
Markdown==3.3.7
matplotlib==3.5.2
numpy==1.23.0
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytesseract==0.3.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
sb3-contrib==1.5.0
six==1.16.0
stable-baselines3==1.5.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tesseract==0.1.3
torch==1.12.0
torch-tb-profiler==0.4.0
typing_extensions==4.3.0
urllib3==1.26.9
Werkzeug==2.1.2
wincertstore==0.2
wrapt==1.14.1
zipp==3.8.0
